Clean the benchmark, the tests and the examples
