+++ examples/normal_distribution.rs
@@ -1,8 +1,8 @@
-use sgpu_compute::{blocking::GpuCompute, StageDesc};
+use sgpu_compute::prelude::*;
 
 fn main() {
     let gpu = GpuCompute::new();
-    let mut pipeline = gpu.gen_pipeline::<[f32; 10], u32, [f32; 10], 1>(
+    let mut pipeline = gpu.gen_pipeline(
         None,
         [StageDesc {
             name: Some("norm"),
@@ -10,11 +10,8 @@ fn main() {
             entrypoint: "main",
         }],
     );
+    let input: [f32; 100] = std::array::from_fn(|i| i as f32 / 100.0);
     pipeline.write_uniform(&32768);
-    let result = pipeline.run_blocking(
-        &[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0],
-        [(10, 1, 1)],
-        |vals| *vals,
-    );
+    let result: [f32; 100] = pipeline.run(&input, [(10, 1, 1)], |vals| *vals);
     println!("{:?}", result);
 }
